Improves naming of files for UINavigationController specific code, plus other cleanups
