+++ Succinct/UINavigationController/UINavigationController+UILabel.swift
@@ -2,11 +2,11 @@ import UIKit
 
 extension UINavigationController {
     func hasLabelInNavigationBar(withExactText searchText: String) -> Bool? {
-        if let titleView = self.navigationItem.titleView, titleView.hasLabel(withExactText: searchText) {
+        if let titleView = navigationItem.titleView, titleView.hasLabel(withExactText: searchText) {
             return true
         }
 
-        if self.view.hasLabel(withExactText: searchText) {
+        if view.hasLabel(withExactText: searchText) {
             return true
         }
 
@@ -14,11 +14,11 @@ extension UINavigationController {
     }
 
     func hasLabelInNavigationBar(containingText searchText: String) -> Bool? {
-        if let titleView = self.navigationItem.titleView, titleView.hasLabel(containingText: searchText) {
+        if let titleView = navigationItem.titleView, titleView.hasLabel(containingText: searchText) {
             return true
         }
 
-        if self.view.hasLabel(containingText: searchText) {
+        if view.hasLabel(containingText: searchText) {
             return true
         }
 
@@ -26,14 +26,14 @@ extension UINavigationController {
     }
 
     func findLabelInNavigationBar(withExactText searchText: String) -> UILabel? {
-        if  let titleView = self.navigationItem.titleView,
+        if  let titleView = navigationItem.titleView,
             let labelInTitleView = titleView.findLabel(withExactText: searchText) {
 
             return labelInTitleView
         }
 
 
-        if let labelInNavCtrlHierarchy = self.view.findLabel(withExactText: searchText) {
+        if let labelInNavCtrlHierarchy = view.findLabel(withExactText: searchText) {
             return labelInNavCtrlHierarchy
         }
 
@@ -41,14 +41,14 @@ extension UINavigationController {
     }
 
     func findLabelInNavigationBar(containingText searchText: String) -> UILabel? {
-        if  let titleView = self.navigationItem.titleView,
+        if  let titleView = navigationItem.titleView,
             let labelInTitleView = titleView.findLabel(containingText: searchText) {
 
             return labelInTitleView
         }
 
 
-        if let labelInNavCtrlHierarchy = self.view.findLabel(containingText: searchText) {
+        if let labelInNavCtrlHierarchy = view.findLabel(containingText: searchText) {
             return labelInNavCtrlHierarchy
         }
